// src/lib/counterparty/api.d.ts
export interface Balance {
    asset: string
    qty: number
    qty_normalized: string
    description: string | null
    name: string
    divisible: boolean
    locked: boolean
    utxo: string | null
    address: string | null
    utxo_address: string | null
}